pyproject.toml
[tool.poetry]
name = "telegram-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "telegram_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
python-telegram-bot = "^20.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
